+++ dart/dynamics/detail/GenericJoint.hpp
@@ -1467,18 +1467,6 @@ void GenericJoint<ConfigSpaceT>::setRestPosition(size_t index, double q0)
     return;
   }
 
-  if (Base::mAspectProperties.mPositionLowerLimits[index] > q0
-      || Base::mAspectProperties.mPositionUpperLimits[index] < q0)
-  {
-    dtwarn << "[GenericJoint::setRestPosition] Value of _q0 [" << q0
-           << "], is out of the limit range ["
-           << Base::mAspectProperties.mPositionLowerLimits[index] << ", "
-           << Base::mAspectProperties.mPositionUpperLimits[index]
-           << "] for index [" << index << "] of Joint [" << this->getName()
-           << "].\n";
-    return;
-  }
-
   GenericJoint_SET_IF_DIFFERENT(mRestPositions[index], q0);
 }